rename sdk to core
